Co-creation-projects/TenSon19920106-PBC-Analysis/requirements.txt
hello-agents[all]>=0.1.0
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=7.0.0
requests>=2.28.0
streamlit>=1.28.0
